+++ src/boards/serializers.py
@@ -11,6 +11,10 @@ class BoardSerializer(ValidateModelMixin, serializers.ModelSerializer):
     """
     DRF serializer for 'boards.Board' Django model.
     """
+    pipelines = serializers.HyperlinkedIdentityField(
+        view_name='api:board-pipelines',
+    )
+
     class Meta:
         model = models.Board
         fields = '__all__'